requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Deep Learning
tensorflow>=2.13.0
keras>=2.13.0

# Data Visualization
plotly>=5.14.0
streamlit>=1.25.0

# Configuration
pyyaml>=6.0

# Web Framework (API)
flask>=2.3.0

# Utilities
python-dateutil>=2.8.0

# Logging
colorlog>=6.7.0

